Parser/Interpreter: Initial support for statement blocks
This is the initial work on supporting statement blocks as a special
parser mode that allows for friendlier syntax with less parentheses. The
constraint in this mode is that expressions are allowed to leave off
parentheses if they start with an identifier. Whitespace inside
statement blocks is newline sensitive, unlike in the regular
parenthesized list mode.