Parser/Interpreter: Support for top-level statements
Because all lists are inherently expressions anyways, this allows for
programs to implicitly be statement blocks. This allows for not needing
parentheses for top level statements, and all users of this language
will enjoy one less level of parenthesization.

For example, this is a valid program now:

```
def x 4
set! x (+ x 2)
display x; (newline)
```

Note that in the example above `(newline)` requires parentheses because
otherwise it will evaluate `newline` the symbol as an expression and
yield the function pointer itself. To invoke an unparenthesized
expression that has no arguments, you must surround it with parentheses.
Also note that semicolons may be used to include multiple statements on
one line.