Runtime: Add aliases for common .NET types
This is a precursor to supporting strong typing (at least, runtime
strong type checks for now). The following types are now defined as
globals, which can be shadowed in child scopes: i8, u8, i16, u16,
i32, u32, i64, u64, f32, f64, dec, str, obj, bool, and void.

The `str` function which was a convenience method to feel more
Clojure-like has been at least temporarily renamed to `->string`, due
to the `str` alias now pointing to the type System.String. We might
be able to bring this conversion function back by having `str` return
an object that can both be implicitly convertible to System.String
as well as being invocable as a function to convert its argument to
a string. This could also then be done for the other type aliases as
well. But for now we're deferring solving that, as this function or the
.NET .ToString function can suffice.