Interpreter: Add return type syntax and basic runtime type checking
This is the beginning of work to support basic runtime type checking by
checking that the return type of a method matches the specified return
type in the signature, if there is one provided.

This does not yet validate parameters or provide type inference or
anything like that; it's just the first step down this path.