Interpreter: Support tail call support for `and` and `or`
This adds tail call support for `and` and `or` per R7RS-small 3.5.
Additionally, some logic in both methods was fixed per 4.2.1, such as
behavior if no arguments are provided, and returning non-false results
instead of exactly #t or #f.

In order to better test this, a new RuntimeOptions parameter has been
added to the runtime that allows for specifying a maximum stack depth.
This is so that we can simulate a very small stack and not have to run
full, exhaustive stack overflow tests.

By testing this, it exposed that the tail-call optimization logic as
previously implemented in Lillisp was incomplete. While the .NET stack
did not get exhausted, it was creating a new environment/scope for each
tail call. This has been fixed by popping and merging the stack for each
tail call after its arguments have been resolved. While the tests all
pass, I'm still not 100% sure this is a rock-solid solution.