Interpreter: Better support for number testing
This improves the logic for detecting if a number is an integer, real,
or rational. Several previously-skipped unit tests are now working, and
a few more were added. We currently cannot discriminate between complex
numbers 3+0i and 3+0.0i, as the .NET Complex type only has double
components. Might need to add our own generic Complex type to the code
base to support this.