Runtime: Initial support for Scheme libraries
This refactors the R7RS-small libraries out as automatically-imported
standard libraries, and adds basic support for the `import` form that
takes a list of library names. This is tested via unit tests that tell
the runtime to not automatically import the standard libraries by
default.

Work still remains on several library methods, moving the
Colibri-specific functions into their own libraries (i.e. `(colibri
linq)`), and other forms of `import`. And of course, you can't yet
define a library, so that will need to happen soon as well.