REPL: Much improved prompt
This adds PrettyPrompt to make the prompt a lot friendlier to use, and
solves a few bugs. Keyboard up/down navigation now works, and Ctrl+D is
now supported as an additional exit shortcut. This should also enable us
in the future to support things like completion and hints as well.