Base: Add `boolean=?` function
Note the comment about spec compliance. We are taking a more forgiving
reading of the R7RS-small spec, allowing for passing non-booleans and
returning #f, as well as passing less than two arguments. The spec does
indicate that it requires at least two arguments in the syntax, but then
the description of it seems to contraindicate that requirement.