Base: Add basic `dynamic-wind` support
I'm not an expert at using `dynamic-wind`, but it seems like a good fit
for `try`/`finally`. Note that some implementations do not seem to call
the "out" thunk in the event of a raised exception, but per the spec it
seems like we should. This is functionality that comes "for free" by
using `try`/`finally`. If anyone notices a flaw in this logic, please
let me know!