ProcessContext: Add `exit` function
This adds the `exit` function as one of the last R7RS functions
implemented. This seems odd at first glance, as we already had `exit` as
a special command in the REPL, and `emergency-exit` implemented. But
really this is because it requires unwinding any `dynamic-wind`
windings, and that was only recently implemented.

Note that the spec says that this should not throw an exception, but
that's how we handle `dynamic-wind` via try/finally. So here we've made
the Scheme exception handling methods rethrow an ExitException if it
encounters one and let the Runtime catch that. However, theoretically
some .NET interop code with try/catch could intercept this. I think it's
okay that we're fudging the spec a little on this one since from purely
Scheme code it will not surface to the user as an exception.

Anyone that uses the ColibriRuntime directly will need to implement a
handler for the Exit event, or else `(exit)` will throw an
ExitException.