R5RS: Add `(scheme r5rs)` library and option to import it only
This adds the r5rs standard library with most methods (except
null-environment and scheme-report-environment) as an available library
to import. Of course, this doesn't do much good if r7rs libraries are
already imported, so this adds the ability to run the REPL or
interpreter with the `--r5rs` flag which will skip importing the R7RS
libraries and import R5RS instead.